+++ packages/internal-lib-build/configs/jest/setup-jest.js
@@ -30,12 +30,13 @@ if (typeof global.performance !== 'undefined' && typeof global.performance.mark 
     global.performance.clearMeasures = nodePerformance.clearMeasures.bind(nodePerformance)
 }
 
-// Polyfill TextDecoder for jsdom environment
+// Polyfill TextDecoder/TextEncoder for jsdom environment
 // Required for AWS SDK 3.x's @smithy/core CBOR encoding (used by mrt-utilities)
 if (typeof global.TextDecoder === 'undefined') {
     // eslint-disable-next-line @typescript-eslint/no-var-requires
-    const {TextDecoder} = require('util')
+    const {TextDecoder, TextEncoder} = require('util')
     global.TextDecoder = TextDecoder
+    global.TextEncoder = TextEncoder
 }
 
 // Mock Fetch